Start moving functions from creaetReviewPages
